feat(harness/lua-server): dual-protocol Lua adapter (mirror of harness/ruby-server)
The earlier harness/lua-server/server.lua was MPP-only. With the
resty.pay_kit umbrella + the 11-rule x402 verifier port now in
place, the Lua side gets the same dual-protocol harness adapter
shape Ruby ships: one binary, two settle paths, picked per scenario
by which env namespace the orchestrator sets (or the explicit
PAY_KIT_INTEROP_PROTOCOL hint that the matrix uses when both
X402_INTEROP_* and MPP_INTEROP_* are populated against the same
surfpool fixtures).

- harness/lua-server/server.lua            New dual-protocol adapter.
                                            Reads either MPP_INTEROP_*
                                            or X402_INTEROP_*; configures
                                            resty.pay_kit with the right
                                            operator/signer/secret;
                                            registers a single gate; runs
                                            a raw socket loop that routes
                                            GET /<resource> through
                                            pay_kit.try_payment. Honours
                                            *_INTEROP_SETTLEMENT_HEADER
                                            so per-scenario custom header
                                            names land on the wire.
- harness/lua-server/server.legacy.lua     The previous MPP-only adapter
                                            (370 LOC), preserved as a
                                            reference. Drop after the
                                            matrix's lua x402 step is
                                            green.

- harness/src/implementations.ts            `lua` impl now declares
                                            intents: ["charge",
                                            "x402-exact"], so the matrix
                                            will pair it against x402
                                            scenarios alongside MPP ones.
                                            Label updated to "Lua PayKit
                                            server (dual protocol)".

Local: both modes boot cleanly (verified via PAY_KIT_INTEROP_PROTOCOL
hint smoke tests); 449 tests pass / 0 fail in the Lua suite; luacheck
clean; harness pnpm typecheck clean.

Matrix wiring: with this adapter, MPP_INTEROP_SERVERS=lua +
MPP_INTEROP_INTENTS=charge,x402-exact in the cross-language matrix
will exercise Lua against both protocols end-to-end (mirrors the
Ruby pay-kit-server flow that landed via PR #138). The remaining
piece - adding ruby-style serverIds entries that include `lua` for
the negative-path scenarios (network-mismatch, cross-route-replay,
idempotent-resubmit) - is a focused follow-up once a lua matrix
scenario lands green.
